project 1 completion
Current Implementation:

The nav on left shows which view is active. The menu uses global Menu
and Menu Item components from material-ui library. The nav also is in
it’s own custom component and accepts the current route to decide which
page is active.

We loop through the images and display all uploaded images with the
global Card, Card Title, and Card Media components from the material-ui
library.

if there is an error on upload, an alert appears with the error message
that hides itself after 4 seconds. The alert uses the global snackbar
component form material-ui.

The nav hides itself on smaller screens.

The layout uses css grid and the rest of the global layout is set with
a muiTheme from material-ui library.

The upload button uses the Flat Button global component from
material-ui for a better look and feel.

The progress bar appears when a file is uploading and uses the global
progress bar component form material-ui. It is wrapped in it’s own
component and accept the value parameter.

If you have no images uploaded yet, it will guide you as to how to get
started. If you already have images, the prompt does not display.

As the app is loading, you see a circular progress from the material-ui
component library.

Future Plans:

When you click on favorites from the nav, it would filter to only show
your favorite files

When you click on the image heart, it would get a new fav parameter
tied to it.

On smaller screens, the menu becomes a dropdown menu from the app bar
hamburger.

The columns of the grid list for displaying images would go from 3 to 2
to 1 depending on the screen width.